+++ tls/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added a p256 implementation optimized for the Cortex-M4 CPU with the `p256-cm4` feature.
+
 ## [0.1.1] - 2022-07-26
 ### Fixed
 - Fixed a race condition where the client would fail to send finished before the server connection timeout expired.